Pull out Secp256k1Wallet into separate module
